nssit-api/requirements.txt
alembic==1.4.2
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
click==7.1.2
commonmark==0.9.1
cryptography==3.0
Flask==1.1.2
flask-marshmallow==0.13.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
Padding==0.5
pycparser==2.20
pycrypto==2.6.1
pycryptodome
Pygments==2.6.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.18
typing-extensions==3.7.4.2
urllib3==1.25.10
Werkzeug==1.0.1
psycopg2
#gunicorn